build: move release-only scripts into scripts/release/
version.py, update_version.py, update_patch_version.py, and
version_files.py are only ever used for cutting a release. Group them
under scripts/release/ to make that boundary explicit.

repo_targets.py and repo.toml stay put: repo_targets.py is also
imported by scripts/griffe_check.py for the public-symbols-check tox
env (unrelated to releases), so repo.toml isn't exclusively a release
file.

The 3 moved entry points import repo_targets from the parent scripts/
directory, which requires an explicit sys.path insertion since Python
only auto-adds a directly-run script's own directory to sys.path.
Added matching E402 per-file-ignores in pyproject.toml, mirroring the
existing exception for opentelemetry-sdk/tests/_configuration/test_models.py.
